backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
gunicorn==20.1.0
gunicorn==20.1.0
gunicorn
psycopg2-binary
gevent